Engine primitives (wavesdb: Snapshot/SSTablesSince/Checkpoint/Restore), single-node logical
backup (full + re-shard + partial), and the cluster layer: coordination (3a), physical
incrementals (3b), bootstrap-restore + clone (3c) + transport-completeness, lifecycle/GC (3d),
destination override + S3 config (3e), admin Backup UI (3f). All phases TDD'd + two-stage reviewed.
HLC consistency cut (3a.1) tracked as follow-up.

+++ internal/recordstore/encode.go
@@ -90,7 +90,9 @@ func parseTTLKey(k []byte) (namespace string, userKey []byte, ok bool) {
 	}
 	rest := k[9:] // skip sentinel + 8-byte bucket start
 	nsLen, n := binary.Uvarint(rest)
-	if n <= 0 || int(nsLen) > len(rest)-n {
+	// Unsigned compare (no lossy int() cast): a length in (2^63, 2^64) would go
+	// negative as an int and slip past a signed guard, panicking the slice below.
+	if n <= 0 || nsLen > uint64(len(rest)-n) {
 		return "", nil, false
 	}
 	return string(rest[n : n+int(nsLen)]), rest[n+int(nsLen):], true
